Version 1.0. / Add cp option to disable supplied remote
Configured for Board version 1.3

+++ fluxcapacitor-A10001986/fc_audio.cpp
@@ -126,7 +126,7 @@ void audio_setup()
     audioLogger = &Serial;
     #endif
 
-    useVKnob = ((int)atoi(settings.useVknob) > 0);
+    useVKnob = (atoi(settings.useVknob) > 0);
 
     // Set resolution for volume pot
     analogReadResolution(POT_RESOLUTION);
